requirements.txt
# Path: /home/chanakya/sound_classification/requirements.txt

numpy==1.24.3
pandas==2.0.3
librosa==0.10.1
scikit-learn==1.3.0
tensorflow==2.13.0
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
pydantic-settings==2.0.3
streamlit==1.27.0
plotly==5.17.0
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
pytest==7.4.2

